Name placements support for abbreviated names, fix place prioritization.
This merge is a little garbled, but basically this uses mapnik
"Placement" support to display either the original name of a highway/
city/locality/etc. from the OSM data, or a version automatically
abbreviated by osm2pgsql (if patched appropriately), or a short
name from the short_name tag supplied by users in OSM data -- this
one doesn't get used very much.  This assumes the value in the
shortest_name column is always shorter or same as the short_name
column, which is shorter or same as "name".  The former two columns
either come from OSM data or are auto-generated if the code knows
some of the abbreviatable words present in the full name.  Mapnik
will chose the longest form that fits at given zoom and density of
other labels in the current tile.

Also fix the order place=* nodes are returned from postgis to
mapnik to make sure the important places get displayed and the less
important ones may be omitted if there isn't enough space.