Practical area:highway rendering.
This assumes osm2pgsql translates area:highway=x to highway=x + area=yes
during import since these should have mostly the same rendering even if
they're not semantically equivalent.

Most highway areas rendering is deferred to z16 or z17, the assumption
being that linear features for these highways are present and visible on
lower zoom levels.

Some of the changes here are pretty subjective, this includes the move
of path, cycleway and footway rendering to on top of other area types so
as to accomodate crossings and similar, but under their respective areas.
This is motivated by the fact that as linear features they're rendered
as thin lines and as areas their rendering is completely different which
would otherwise make them stand out when both linear and polygonal
representations are there.